# Admin/bin/Debug/netcoreapp3.1/Docs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_GetOrganizations",
        "parameters": [
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortKey8"
            },
            "x-position": 1
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Order"
            },
            "x-position": 2
          },
          {
            "name": "Name",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "StartDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "StartDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "EndDateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "EndDateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "IncludingEndDateNotSet",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 8
          },
          {
            "name": "IsValid",
            "in": "query",
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 2147483647.0,
              "minimum": 1.0,
              "nullable": true
            },
            "x-position": 10
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000.0,
              "minimum": 1.0
            },
            "x-position": 11
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfOrganization"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_PostOrganization",
        "requestBody": {
          "x-name": "organization",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Organization"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    },
    "/api/organizations/mine": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_GetOrganization",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_PutOrganization",
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutOrganizationParam"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "UserAdmin"
            ]
          }
        ]
      }
    },
    "/api/organizations/{code}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_GetOrganization2",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_PutOrganization2",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "param",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutOrganizationAdminParam"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Organizations"
        ],
        "operationId": "Organizations_DeleteOrganization",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": [
              "SuperAdmin"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Organization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "address",
          "phone",
          "url",
          "adminPhone",
          "adminMail",
          "distributionServerIp",
          "startDate",
          "isValid"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "phone": {
            "type": "string",
            "format": "phone",
            "maxLength": 11,
            "minLength": 10
          },
          "url": {
            "type": "string",
            "format": "uri",
            "minLength": 1
          },
          "adminPhone": {
            "type": "string",
            "format": "phone",
            "maxLength": 11,
            "minLength": 10
          },
          "adminMail": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "distributionServerIp": {
            "type": "string",
            "minLength": 1,
            "pattern": "((2(5[0-5]|[0-4]\\d))|[0-1]?\\d{1,2})(\\.((2(5[0-5]|[0-4]\\d))|[0-1]?\\d{1,2})){3}"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isValid": {
            "type": "boolean"
          }
        }
      },
...
      "Order": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Asc",
          "Desc"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "PaginatedResponseOfOrganization": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Organization"
            }
          }
        }
      },
      "SortKey8": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Code",
          "Name",
          "Address",
          "Phone",
          "AdminPhone",
          "AdminMail",
          "StartDate",
          "EndDate",
          "Url",
          "IsValid"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9
        ]
      },
      "PutOrganizationParam": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "address",
          "phone",
          "url",
          "adminPhone",
          "adminMail"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "phone": {
            "type": "string",
            "format": "phone",
            "maxLength": 11,
            "minLength": 10
          },
          "url": {
            "type": "string",
            "format": "uri",
            "minLength": 1
          },
          "adminPhone": {
            "type": "string",
            "format": "phone",
            "maxLength": 11,
            "minLength": 10
          },
          "adminMail": {
            "type": "string",
            "format": "email",
            "minLength": 1
          }
        }
      },
      "PutOrganizationAdminParam": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PutOrganizationParam"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "startDate",
              "isValid",
              "distributionServerIp"
            ],
            "properties": {
              "startDate": {
                "type": "string",
                "format": "date-time",
                "minLength": 1
              },
              "endDate": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "isValid": {
                "type": "boolean"
              },
              "distributionServerIp": {
                "type": "string",
                "minLength": 1,
                "pattern": "((2(5[0-5]|[0-4]\\d))|[0-1]?\\d{1,2})(\\.((2(5[0-5]|[0-4]\\d))|[0-1]?\\d{1,2})){3}"
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ],
...
}